+++ platforms/fabric/src/main/java/com/dfsek/terra/fabric/mixin/implementations/chunk/ChunkRegionMixin.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.fabric.mixin.world;
+package com.dfsek.terra.fabric.mixin.implementations.chunk;
 
 import com.dfsek.terra.api.platform.block.Block;
 import com.dfsek.terra.api.platform.block.BlockData;
