+++ platforms/fabric/src/main/java/com/dfsek/terra/fabric/mixin/implementations/inventory/item/ItemMixin.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.fabric.mixin.inventory;
+package com.dfsek.terra.fabric.mixin.implementations.inventory.item;
 
 import com.dfsek.terra.api.platform.inventory.ItemStack;
 import net.minecraft.item.Item;
