+++ platforms/fabric/src/main/java/com/dfsek/terra/fabric/mixin/implementations/world/ServerWorldMixin.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.fabric.mixin.world;
+package com.dfsek.terra.fabric.mixin.implementations.world;
 
 import com.dfsek.terra.api.math.vector.Location;
 import com.dfsek.terra.api.platform.block.Block;
@@ -13,40 +13,33 @@ import com.dfsek.terra.fabric.world.block.FabricBlock;
 import com.dfsek.terra.fabric.world.generator.FabricChunkGeneratorWrapper;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.ChunkRegion;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Implements;
 import org.spongepowered.asm.mixin.Interface;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 
-@Mixin(ChunkRegion.class)
+@Mixin(ServerWorld.class)
 @Implements(@Interface(iface = World.class, prefix = "terra$"))
-public abstract class ChunkRegionWorldMixin {
-    @Shadow
-    @Final
-    private ServerWorld world;
-
+public abstract class ServerWorldMixin {
     public int terra$getMaxHeight() {
-        return ((ChunkRegion) (Object) this).getDimensionHeight();
+        return ((ServerWorld) (Object) this).getDimensionHeight();
     }
 
     public ChunkGenerator terra$getGenerator() {
-        return (ChunkGenerator) ((ChunkRegion) (Object) this).toServerWorld().getChunkManager().getChunkGenerator();
+        return (ChunkGenerator) ((ServerWorld) (Object) this).getChunkManager().getChunkGenerator();
     }
 
     public Chunk terra$getChunkAt(int x, int z) {
-        return (Chunk) ((ChunkRegion) (Object) this).getChunk(x, z);
+        return (Chunk) ((ServerWorld) (Object) this).getChunk(x, z);
     }
 
     public Block terra$getBlockAt(int x, int y, int z) {
-        return new FabricBlock(new BlockPos(x, y, z), ((ChunkRegion) (Object) this));
+        return new FabricBlock(new BlockPos(x, y, z), ((ServerWorld) (Object) this));
     }
 
     public Entity terra$spawnEntity(Location location, EntityType entityType) {
-        net.minecraft.entity.Entity entity = ((net.minecraft.entity.EntityType<?>) entityType).create(((ChunkRegion) (Object) this).toServerWorld());
+        net.minecraft.entity.Entity entity = ((net.minecraft.entity.EntityType<?>) entityType).create(((ServerWorld) (Object) this));
         entity.setPos(location.getX(), location.getY(), location.getZ());
-        ((ChunkRegion) (Object) this).spawnEntity(entity);
+        ((ServerWorld) (Object) this).spawnEntity(entity);
         return (Entity) entity;
     }
 
@@ -64,20 +57,5 @@ public abstract class ChunkRegionWorldMixin {
 
     public TerraChunkGenerator terra$getTerraGenerator() {
         return ((FabricChunkGeneratorWrapper) terra$getGenerator()).getHandle();
-    }
-
-    /**
-     * We need regions delegating to the same world
-     * to have the same hashcode. This
-     * minimizes cache misses.
-     * <p>
-     * This is sort of jank, but shouldn't(tm)
-     * break any other mods, unless they're doing
-     * something they really shouldn't, since
-     * ChunkRegions are not supposed to persist.
-     */
-    @Override
-    public int hashCode() {
-        return world.hashCode();
     }
 }
