+++ platforms/fabric/src/main/resources/terra.mixins.json
@@ -4,25 +4,25 @@
   "package": "com.dfsek.terra.fabric.mixin",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "MixinGeneratorOptions",
-    "block.BlockEntityMixin",
-    "block.LootableContainerBlockEntityMixin",
-    "entity.EntityMixin",
-    "entity.EntityTypeMixin",
-    "entity.PlayerEntityMixin",
-    "entity.ServerCommandSourceMixin",
-    "inventory.EnchantmentMixin",
-    "inventory.ItemMixin",
-    "inventory.ItemStackDamageableMixin",
-    "inventory.ItemStackMetaMixin",
-    "inventory.ItemStackMixin",
-    "world.BiomeMixin",
-    "world.ChunkGeneratorMixin",
-    "world.ChunkRegionMixin",
-    "world.ChunkRegionWorldMixin",
-    "world.ProtoChunkMixin",
-    "world.ServerWorldMixin",
-    "world.WorldChunkMixin"
+    "GeneratorOptionsMixin",
+    "implementations.block.BlockEntityMixin",
+    "implementations.block.state.LootableContainerBlockEntityMixin",
+    "implementations.entity.EntityMixin",
+    "implementations.entity.EntityTypeMixin",
+    "implementations.entity.PlayerEntityMixin",
+    "implementations.entity.ServerCommandSourceMixin",
+    "implementations.inventory.meta.EnchantmentMixin",
+    "implementations.inventory.item.ItemMixin",
+    "implementations.inventory.meta.ItemStackDamageableMixin",
+    "implementations.inventory.meta.ItemStackMetaMixin",
+    "implementations.inventory.item.ItemStackMixin",
+    "implementations.BiomeMixin",
+    "implementations.ChunkGeneratorMixin",
+    "implementations.chunk.ChunkRegionMixin",
+    "implementations.world.ChunkRegionMixin",
+    "implementations.chunk.data.ProtoChunkMixin",
+    "implementations.world.ServerWorldMixin",
+    "implementations.chunk.WorldChunkMixin"
   ],
   "client": [
     "GeneratorTypeAccessor"
